Refactor CentralContainer and extract Sidebar component.
