+++ new-log-viewer/src/components/CentralContainer/Sidebar.css
@@ -1,0 +1,3 @@
+.sidebar-tabs-container {
+    display: flex;
+}
